+++ R/climatekit-package.R
@@ -3,6 +3,5 @@
 
 ## usethis namespace: start
 #' @importFrom stats pnorm qnorm
-#' @importFrom tools R_user_dir
 ## usethis namespace: end
 NULL